# specs/openapi.yaml
openapi: 3.1.0
info:
  title: What's For Supper API
  description: API for the What's For Supper PWA, including recipe discovery, planning, and management.
  version: 0.1.0
servers:
  - url: /
    description: Unified API (same origin)
  - url: http://localhost:9001
    description: Local direct API
security:
  - HearthSecret: []
  - HearthToken: []
components:
  securitySchemes:
    FamilyMemberId:
      type: apiKey
      in: header
      name: X-Family-Member-Id
    HearthSecret:
      type: apiKey
      in: header
      name: X-Hearth-Secret
    HearthToken:
      type: http
      scheme: bearer
      bearerFormat: HMAC
  schemas:
...
    RecipeDto:
      type: object
      required: [id, name, imageUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        imageUrl: { type: [string, 'null'], example: 'https://images.unsplash.com/photo-1546069901-ba9599a7e63c' }

        totalTime: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        dietaryProfile:
          nullable: true
          oneOf:
            - { $ref: '#/components/schemas/RecipeDietaryProfileDto' }
            - { type: 'null' }
        difficulty: { type: [string, 'null'] }
        images:
          type: array
          items: { type: integer }
        rating: { type: integer }
        isVegetarian: { type: boolean }
        isHealthyChoice: { type: boolean }
        isDiscoverable: { type: boolean }
        notes: { type: [string, 'null'] }
        sourceUrl: { type: [string, 'null'] }
        addedBy: { type: [string, 'null'], format: uuid }
        ingredients:
          type: [array, 'null']
          items: { type: string }
        recipeInstructions:
          type: [array, 'null']
          items: {}
          description: 'Parsed recipe instructions (string array or HowToStep objects)'
        createdAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
    RecipeListResponse:
      type: object
      required: [updatedAt, recipes, pagination]
      properties:
        updatedAt: { type: string, format: date-time }
        recipes:
          type: array
          items: { $ref: '#/components/schemas/RecipeDto' }
        pagination:
          $ref: '#/components/schemas/PaginationDto'
    RecipeLibrarySummaryDto:
      type: object
      required: [total, neverCooked, ratings]
      properties:
        total: { type: integer, description: 'Count of all non-soft-deleted recipes.' }
        neverCooked: { type: integer, description: 'Count where lastCookedDate IS NULL and deletedAt IS NULL.' }
        ratings:
          type: object
          required: [love, like, dislike, unrated]
          properties:
            love: { type: integer, description: 'rating == 3' }
            like: { type: integer, description: 'rating == 2' }
            dislike: { type: integer, description: 'rating == 1' }
            unrated: { type: integer, description: 'rating == 0' }
    RecipeDetailResponse:
      type: object
      required: [updatedAt, recipe]
      properties:
        updatedAt: { type: string, format: date-time }
        recipe: { $ref: '#/components/schemas/RecipeDto' }
    PaginationDto:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    RecipeSearchFiltersDto:
      type: object
      properties:
        newRecipes: { type: [boolean, 'null'] }
        neverCooked: { type: [boolean, 'null'] }
        familyFavorite: { type: [boolean, 'null'] }
        quickOnly: { type: [boolean, 'null'] }
        notCookedInLongTime: { type: [boolean, 'null'] }
        discoverableOnly: { type: [boolean, 'null'] }
    RecipeSearchReasonDto:
      type: object
      required: [source, label]
      properties:
        source:
          type: string
          enum:
            - name-match
            - notes-match
            - rating-boost
            - vote-boost
            - planner-fit
            - inventory-fit
            - semantic-match
        label: { type: string }
    RecipeSearchResultDto:
      type: object
      required: [id, rating, isDiscoverable, reasons]
      properties:
        id: { type: string, format: uuid }
        name: { type: [string, 'null'] }
        imageUrl: { type: [string, 'null'] }
        totalTime: { type: [string, 'null'] }
        difficulty: { type: [string, 'null'] }
        rating: { type: integer }
        isDiscoverable: { type: boolean }
        notes: { type: [string, 'null'] }
        reasons:
          type: array
          items: { $ref: '#/components/schemas/RecipeSearchReasonDto' }
        plannerFitNote: { type: [string, 'null'] }
    RecipeSearchRequestDto:
      type: object
      required: [query]
      properties:
        query: { type: string }
        mode: { type: string, enum: [standard, agent], default: standard }
        similarToRecipeId: { type: [string, 'null'], format: uuid }
        pantrySnapshotId: { type: [string, 'null'], format: uuid }
        weekOffset: { type: [integer, 'null'] }
        dayIndex: { type: [integer, 'null'] }
        limit: { type: integer, minimum: 1, maximum: 5, default: 5 }
        filters: { $ref: '#/components/schemas/RecipeSearchFiltersDto' }
    RecipeSearchResponseDto:
      type: object
      required: [results, appliedFilters, searchMode, resultPath]
      properties:
        topPick:
          oneOf:
            - { $ref: '#/components/schemas/RecipeSearchResultDto' }
            - { type: 'null' }
        results:
          type: array
          items: { $ref: '#/components/schemas/RecipeSearchResultDto' }
        appliedFilters: { $ref: '#/components/schemas/RecipeSearchFiltersDto' }
        searchMode:
          type: string
          enum: [standard, agent, similar, pantry-assisted]
        resultPath:
          type: string
          enum: [lexical-only, hybrid, fallback-lexical]
...
    FopFlagsDto:
      type: object
      required: [highInSaturatedFat, highInSugars, highInSodium]
      properties:
        highInSaturatedFat: { type: boolean }
        highInSugars:       { type: boolean }
        highInSodium:       { type: boolean }
...
    RecipeDietaryProfileDto:
      type: object
      required: [primaryFoodGroup, secondaryFoodGroups, proteinSource, cuisineType,
                 mealTypes, primaryMealType, wholeGrainConfident, confidence, source]
      properties:
        primaryFoodGroup:    { type: string }
        secondaryFoodGroups: { type: array, items: { type: string } }
        proteinSource:       { type: string }
        cuisineType:         { type: string }
        mealTypes:           { type: array, items: { type: string } }
        primaryMealType:     { type: string }
        wholeGrainConfident: { type: boolean }
        confidence:          { type: number }
        source:              { type: string }
        fopFlags:
          nullable: true
          oneOf:
            - { $ref: '#/components/schemas/FopFlagsDto' }
            - { type: 'null' }
...
    UpdateRecipeDto:
      type: object
      properties:
        notes: { type: [string, 'null'] }
        rating: { type: [integer, 'null'], minimum: 0, maximum: 3 }
        isDiscoverable: { type: [boolean, 'null'] }
    DescribeRecipeDto:
      type: object
      required: [name]
      properties:
        name: { type: string, description: 'Short name for the recipe (e.g. "Our family spaghetti")' }
        description: { type: [string, 'null'], description: 'Free-text description used to synthesize the full recipe via AI' }
    CaptureUrlDto:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri, description: 'The recipe URL to capture' }
        notes: { type: [string, 'null'] }
        rating: { type: [integer, 'null'], minimum: 0, maximum: 3 }
    RecipeStatusDto:
      type: object
      required: [id, status, imageCount, isSynthesized]
      properties:
        id: { type: string, format: uuid }
        name: { type: [string, 'null'] }
        status: { type: string, enum: [pending, ready], description: '"pending" while synthesis is in progress; "ready" once Name is set and either ImageCount > 0 (photo-upload path) or IsSynthesized = true (describe path)' }
        imageCount: { type: integer }
        isSynthesized: { type: boolean, description: 'true once the goto-synthesis workflow has completed successfully' }
...
    RecipeImportTriggerResponseDto:
      type: object
      required: [importId]
      properties:
        importId: { type: string, format: uuid }
    RecipeImportStatusResponseDto:
      type: object
      required: [status]
      properties:
        status: { type: string }
        errorMessage: { type: [string, 'null'] }
...
    RecipeTrashItemDto:
      type: object
      required: [id, deletedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: [string, 'null'] }
        imageUrl: { type: [string, 'null'] }
        deletedAt: { type: string, format: date-time }
        deletedBy: { type: [string, 'null'], format: uuid }
    RecipeTrashListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RecipeTrashItemDto' }
    RecipePurgeResponse:
      type: object
      required: [purged]
      properties:
        purged: { type: boolean }
...
paths:
...
  /api/recipes:
    get:
      summary: List recipes
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [explore]
          description: |
            Sort order for recipes. When "explore" is specified, recipes are ordered by
            lastCookedDate ASC NULLS FIRST (never-cooked first, then oldest-cooked first).
            When absent, default ordering (newest-created first) is applied.
            Returns HTTP 400 for unrecognised values.
        - name: discoverableOnly
          in: query
          schema:
            type: boolean
            default: false
          description: |
            When true, only returns recipes marked as discoverable (high family interest).
            When false (default), returns all non-soft-deleted recipes.
      responses:
        '200':
          description: OK (Not wrapped in data envelope)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeListResponse' }
        '400':
          description: Invalid order parameter value
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
    post:
      summary: Create a recipe
      security: [{ FamilyMemberId: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 0, maximum: 3, default: 0 }
                finishedDishImageIndex: { type: integer, default: -1 }
                notes: { type: [string, 'null'] }
                images:
                  type: array
                  items:
                    type: string
                    format: binary

      responses:
        '202':
          description: Accepted (Import workflow triggered)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
              example:
                data:
                  id: '550e8400-e29b-41d4-a716-446655440003'
  /api/recipes/library-summary:
    get:
      summary: Get recipe library health counts
      description: |
        Returns a lightweight summary of the recipe library for use in the depth indicator.
        All counts exclude soft-deleted recipes.
      security: [{ FamilyMemberId: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/RecipeLibrarySummaryDto'
              example:
                data:
                  total: 42
                  neverCooked: 12
                  ratings:
                    love: 8
                    like: 15
                    dislike: 3
                    unrated: 16
  /api/recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get recipe detail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeDetailResponse' }
    patch:
      summary: Update a recipe
      security: [{ FamilyMemberId: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRecipeDto' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
    delete:
      summary: Soft-delete a recipe (returns updated recipe with deletedAt set)
      security: [{ FamilyMemberId: [] }]
      responses:
        '200':
          description: Soft-deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeDetailResponse' }
        '409':
          description: Recipe is assigned to an active planner slot
          content:
            application/json:
              schema:
                type: object
                required: [errorCode, message, assignedDays]
                properties:
                  errorCode: { type: string, enum: [RECIPE_ASSIGNED_TO_PLANNER] }
                  message: { type: string }
                  assignedDays:
                    type: array
                    items: { type: string, format: date }
  /api/recipes/trash:
    get:
      summary: List soft-deleted recipes (Recycle Bin)
      security: [{ FamilyMemberId: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/RecipeTrashListResponse' }
  /api/recipes/{id}/restore:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Restore a soft-deleted recipe
      security: [{ FamilyMemberId: [] }]
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeDetailResponse' }
  /api/recipes/{id}/purge:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Hard-delete (purge) a soft-deleted recipe from the Recycle Bin
      security: [{ FamilyMemberId: [] }]
      parameters:
        - name: X-Elevated-Pin
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Purged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/RecipePurgeResponse' }
        '403':
          description: Missing or incorrect elevated PIN
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode: { type: string }
                  message: { type: string }
        '409':
          description: Recipe is not in trash (deleted_at is null)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode: { type: string }
                  message: { type: string }
        '503':
          description: Elevated PIN not configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode: { type: string, enum: [PIN_NOT_CONFIGURED] }
                  message: { type: string }
  /api/recipes/{recipeId}/original/{photoIndex}:
    parameters:
      - name: recipeId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: photoIndex
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: Get original recipe image by index
      operationId: getRecipeOriginalImage
      tags: [Recipes]
      responses:
        '200':
          description: Raw image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Recipe or image not found
  /api/recipes/{id}/hero:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get recipe hero image
      responses:
        '200':
          description: Image binary data
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
...
  /api/recipes/search:
    post:
      summary: Search recipes
      operationId: searchRecipes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecipeSearchRequestDto' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/RecipeSearchResponseDto' }
              example:
                data:
                  topPick: null
                  results: []
                  appliedFilters: {}
                  searchMode: standard
                  resultPath: lexical-only
...
  /api/recipes/{id}/import:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Trigger a manual recipe import
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeImportTriggerResponseDto' }
    get:
      summary: Check status of a recipe import
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipeImportStatusResponseDto' }
...
  /api/recipes/describe:
    post:
      security: [{ FamilyMemberId: [] }]
      summary: Create a stub recipe from a text description and trigger AI synthesis
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DescribeRecipeDto' }
            example:
              name: 'Our family spaghetti'
              description: 'Homemade tomato sauce with meatballs, slow-cooked for two hours'
      responses:
        '202':
          description: Accepted — stub recipe created; synthesis workflow triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/RecipeDto' }
              example:
                data:
                  id: '660e8400-e29b-41d4-a716-446655440020'
                  name: 'Our family spaghetti'
                  description: 'Homemade tomato sauce with meatballs, slow-cooked for two hours'
                  imageUrl: null
                  createdAt: '2026-04-30T12:00:00Z'
  /api/recipes/capture-url:
    post:
      summary: Capture a recipe from a URL
      security: [{ FamilyMemberId: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CaptureUrlDto' }
      responses:
        '202':
          description: Accepted (Capture workflow triggered)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
  /api/recipes/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get the synthesis status of a recipe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/RecipeStatusDto' }
              example:
                data:
                  id: '660e8400-e29b-41d4-a716-446655440020'
                  name: 'Our family spaghetti'
                  status: 'pending'
                  imageCount: 0
        '404':
          description: Recipe not found
...